Add a delay before sending commands
Under some circumstances the firmware might take a little longer to
reach the UEFI shell. If the `ESC` key is send and we are still in the
Boot Manager phase where hotkeys are active, we might end up booting
into the firmware menu instead.

Add a 2 second delay before sending the commands to increase the
probability of actually having reached the shell.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>